+++ htdocs/DailyReport.php
@@ -1,6 +1,7 @@
 <?php
 require_once("config.inc.php");
 require_once($cfg["path_bu"] . "/BUDailyReport.inc.php");
+require_once($cfg["path_ct"] . '/CTCNC.inc.php');
 
 $buDailyReport = new BUDailyReport($this);
 
@@ -9,6 +10,8 @@ if ($_REQUEST['daysAgo']) {
 } else {
     $daysAgo = 1;
 }
+
+$controller = new CTCNC(null, null, null, null, null);
 
 switch ($_REQUEST['action']) {
 
@@ -22,7 +25,17 @@ switch ($_REQUEST['action']) {
         $buDailyReport->prepayOverValue($daysAgo);
         break;
     case 'outstandingIncidents' :
-        $buDailyReport->outstandingIncidents($daysAgo);
+        $roles = [
+            "technical",
+        ];
+        if (!$controller->hasPermissions($roles)) {
+            Header("Location: /NotAllowed.php");
+            exit;
+        }
+
+        $onScreen = isset($_GET['onScreen']);
+
+        $buDailyReport->outstandingIncidents($daysAgo, null, $onScreen);
         break;
     case 'outstandingPriorityFiveIncidents' :
         $buDailyReport->outstandingIncidents($daysAgo, true);